guvenlik acigi giderildi.

+++ src/main/java/tr/edu/ogu/ceng/controller/FacultySupervisorController.java
@@ -4,7 +4,9 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -20,6 +23,7 @@ import tr.edu.ogu.ceng.dto.requests.FacultySupervisorRequestDto;
 import tr.edu.ogu.ceng.dto.responses.FacultySupervisorResponseDto;
 import tr.edu.ogu.ceng.model.FacultySupervisor;
 import tr.edu.ogu.ceng.model.User;
+import tr.edu.ogu.ceng.security.UserPrincipal;
 import tr.edu.ogu.ceng.service.FacultySupervisorService;
 import tr.edu.ogu.ceng.util.PageableUtil;
 
@@ -61,18 +65,33 @@ public class FacultySupervisorController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<FacultySupervisorResponseDto> getFacultySupervisor(@PathVariable(name = "id") long id) {
-		FacultySupervisorResponseDto facultySupervisorResponseDto = facultySupervisorService.getFacultySupervisor(id);
-		return ResponseEntity.ok(facultySupervisorResponseDto);
+	public ResponseEntity<FacultySupervisorResponseDto> getFacultySupervisor(@PathVariable(name = "id") long id, Authentication authentication) {
+	    UserPrincipal currentUser = (UserPrincipal) authentication.getPrincipal();
+	    // İsteği yapan kullanıcı yetkilendirilmiş mi kontrolü
+	    if (!currentUser.getUser().getId().equals(id)) {
+	        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+	    }
+	    FacultySupervisorResponseDto facultySupervisorResponseDto = facultySupervisorService.getFacultySupervisor(id);
+	    return ResponseEntity.ok(facultySupervisorResponseDto);
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Boolean> deleteFacultySupervisor(@PathVariable(name = "id") long id) {
-		return ResponseEntity.ok(facultySupervisorService.deleteFacultySupervisor(id));
+	public ResponseEntity<Boolean> deleteFacultySupervisor(@PathVariable(name = "id") long id, Authentication authentication) {
+	    UserPrincipal currentUser = (UserPrincipal) authentication.getPrincipal();
+	    // İsteği yapan kullanıcı yetkilendirilmiş mi kontrolü
+	    if (!currentUser.getUser().getId().equals(id)) {
+	        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+	    }
+	    return ResponseEntity.ok(facultySupervisorService.deleteFacultySupervisor(id));
 	}
 
-	@GetMapping("/byUserId/{userId}")
-	public FacultySupervisorResponseDto getFacultySupervisorByUserId(@PathVariable Long userId) {
-		return facultySupervisorService.getFacultySupervisorByUserId(userId);
+	@RequestMapping(value = "/byUserId/{userId}", method = RequestMethod.GET)
+	public ResponseEntity<FacultySupervisorResponseDto> getFacultySupervisorByUserId(@PathVariable Long userId, Authentication authentication) {
+	    UserPrincipal currentUser = (UserPrincipal) authentication.getPrincipal();
+	    if (currentUser.getUser().getId()!= userId) {
+	        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+	    }
+	    return ResponseEntity.ok(facultySupervisorService.getFacultySupervisorByUserId(userId));
 	}
+
 }
